requirements.txt
Package                   Version
------------------------- -----------
aiohappyeyeballs          2.6.1
aiohttp                   3.12.15
aiosignal                 1.4.0
altair                    5.5.0
annotated-types           0.7.0
anyio                     4.10.0
attrs                     25.3.0
blinker                   1.9.0
cachetools                5.5.2
certifi                   2025.8.3
charset-normalizer        3.4.3
click                     8.2.1
dataclasses-json          0.6.7
distro                    1.9.0
frozenlist                1.7.0
gitdb                     4.0.12
GitPython                 3.1.45
h11                       0.16.0
httpcore                  1.0.9
httpx                     0.27.2
idna                      3.10
Jinja2                    3.1.6
jiter                     0.10.0
jsonpatch                 1.33
jsonpointer               3.0.0
jsonschema                4.25.1
jsonschema-specifications 2025.9.1
langchain                 0.3.0
langchain-community       0.3.0
langchain-core            0.3.63
langchain-openai          0.2.2
langchain-text-splitters  0.3.8
langsmith                 0.1.147
markdown-it-py            4.0.0
MarkupSafe                3.0.2
marshmallow               3.26.1
mdurl                     0.1.2
multidict                 6.6.4
mypy_extensions           1.1.0
narwhals                  2.5.0
numpy                     1.26.4
openai                    1.47.0
orjson                    3.11.3
packaging                 24.2
pandas                    2.3.2
pillow                    11.3.0
pip                       25.2
propcache                 0.3.2
protobuf                  5.29.5
pyarrow                   21.0.0
pydantic                  2.11.8
pydantic_core             2.33.2
pydantic-settings         2.10.1
pydeck                    0.9.1
Pygments                  2.19.2
python-dateutil           2.9.0.post0
python-dotenv             1.1.1
pytz                      2025.2
PyYAML                    6.0.2
referencing               0.36.2
regex                     2025.9.1
requests                  2.32.5
requests-toolbelt         1.0.0
rich                      13.9.4
rpds-py                   0.27.1
six                       1.17.0
smmap                     5.0.2
sniffio                   1.3.1
SQLAlchemy                2.0.43
streamlit                 1.41.1
tenacity                  8.5.0
tiktoken                  0.11.0
toml                      0.10.2
tornado                   6.5.2
tqdm                      4.67.1
typing_extensions         4.15.0
typing-inspect            0.9.0
typing-inspection         0.4.1
tzdata                    2025.2
urllib3                   2.5.0
yarl                      1.20.1
